requirements.txt
streamlit #For the web application interface
Pillow #For image processing/handling
fpdf #For PDF creation
